apps/rag-processor/pyproject.toml
[project]
name = "rag-processor"
version = "0.1.0"
description = "RAG document processing and embedding service - Optimized CPU version"
requires-python = ">=3.10,<3.13"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    # Google Cloud services - optimized for API usage
    "google-cloud-storage>=2.10.0,<3.0.0",
    "vertexai>=1.38.0", # For embeddings via API
    "google-genai>=1.24.0",
    "pydub==0.25.1",
    "google-cloud-secret-manager>=2.18.1",
    # Database
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    # Document processing - lighter alternatives
    "opencv-python-headless>=4.8.0", # Headless wheel for servers (no GUI/GLX deps)
    "docling>=2.41.0",
    "docling-core>=2.42.0", # For HybridChunker
    "huggingface-hub>=0.20.0", # For model downloads
    "requests>=2.32.0", # HTTP requests for model downloads
    # Web framework
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "httpx>=0.25.0", # HTTP client
    "hf-transfer>=0.1.9",
    "psutil>=5.9.0", # System and process utilities for memory monitoring
]

[project.scripts]
# Infrastructure setup
setup-dev = "scripts.setup_gcp_dev:main"
setup-prod = "scripts.setup_gcp_prod:main"



[dependency-groups]
# 🧠 ML/AI dependencies - only installed in cloud
ml = [
    "torch==2.5.0",
    "torchvision==0.20.0",
    "transformers>=4.53.0",
    "accelerate>=1.8.0",
    "docling>=2.41.0",
    "docling-core>=2.42.0", # For HybridChunker
    "huggingface-hub>=0.20.0",
    "tokenizers>=0.21.0",
    "pillow>=11.3.0",
    "numpy>=1.26.0",
    "tqdm>=4.67.0",
    "ffmpeg-python>=0.2.0",
    "mutagen==1.47.0",
    "opencv-python-headless>=4.8.0",
]
dev = [
    "ruff>=0.12.0",
    "black>=25.1.0", 
    "mypy>=1.16.0",
    "pre-commit>=3.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "types-psycopg2>=2.9.21.20250516",
    "types-requests>=2.32.4.20250611",
]
lint = [
    "ruff>=0.12.0",
    "black>=25.1.0",
    "mypy>=1.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag_processor", "scripts"]

[tool.uv]
dev-dependencies = []

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }


[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
# directories
\.eggs
| \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| build
| dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"
src = ["rag_processor", "scripts"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
]
ignore = [
    "E203", # whitespace before : (conflicts with black)
    "E501", # line too long (handled by black)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

# Handle libraries without type stubs
[[tool.mypy.overrides]]
module = "pgvector.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "vertexai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.cloud.storage.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pypdfium2.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
] 
